requirements.txt
numpy
pandas
matplotlib
torch
transformers
bert_score
# bleu_score
datasets
rouge
torchmetrics
rich
IPython
nltk
sentencepiece
absl-py
rouge_score
bertviz
# ipykernel
ipywidgets
flair
seaborn
spacy
pyLDAvis
sumy
# python -c "import nltk; nltk.download('punkt')"
stanza
bert-extractive-summarizer